QD-3694 update JVM inspections for 2022.2
Most inspections were copied verbatim from the default IDEA profile.

In the starter profile, the JUnit inspection is enabled since it points to tests that were probably expected to be run but were not actually run. All other inspections are not 'vital' for running Qodana on a project for the first time.

In the recommended profile, the Kotlin migrations are disabled, just like in IDEA.

In both profiles, BulkFileAttributesRead is disabled since it has been disabled in IDEA due to its unreliable quick fix.