QD-3713 add selected PHP inspections to starter profile
In Blade Templates, unbalanced parentheses lead to a parser error, so perform this check as early as possible.

Mismatches in PHP int ranges are a classic symptom of code that deviated from its documentation. Enabling this inspection in the starter profile draws attention to this and fits the other similar inspections that are enabled as well.